+++ plugins/markdown/plugin-shiki/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.0.0-rc.106](https://github.com/vuepress/ecosystem/compare/v2.0.0-rc.105...v2.0.0-rc.106) (2025-05-26)
+
+### Features
+
+- add attribute `class="language-"` into `<code>` ([#473](https://github.com/vuepress/ecosystem/issues/473)) ([31f4309](https://github.com/vuepress/ecosystem/commit/31f43098a8a13e51510abffb012d14467e958298))
+
 # [2.0.0-rc.105](https://github.com/vuepress/ecosystem/compare/v2.0.0-rc.104...v2.0.0-rc.105) (2025-05-16)
 
 **Note:** Version bump only for package @vuepress/plugin-shiki